+++ src/components/common/like/Like1.styled.ts
